refactor: split large modules into focused files with barrel re-exports
Split the 4 largest files in the codebase into focused, single-responsibility
modules while preserving all public APIs via barrel re-exports.

- query/finderFns (602 → 37-line barrel + 8 modules): generic createTypedFinder<T>
  factory eliminates 4x duplicated finder boilerplate
- 2d/blueprints/booleanOperations (621 → 133-line barrel + 3 modules): 27 non-null
  assertions eliminated via safe accessors and optional chaining
- topology/shapeHelpers (794 → 46-line barrel + 4 modules): organized by shape
  category (curves, surfaces, solids, utils)
- sketching/Sketcher2d (710 → 695 lines + ellipseUtils): extracted ellipse math,
  added _saveCurveAndAdvance/_resolveRelative/_lastCurve helpers
